hash-service/requirements.txt
flask==2.0.1
werkzeug==2.0.1
opentelemetry-api==1.11.1
opentelemetry-sdk==1.11.1
opentelemetry-exporter-jaeger==1.11.1
opentelemetry-instrumentation-flask==0.30b1
opentelemetry-instrumentation==0.30b1
prometheus-flask-exporter==0.18.7
requests==2.28.1
